% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hamming_distance.R
\name{calculate_hamming_distance}
\alias{calculate_hamming_distance}
\title{Calculates Hamming distance between two vectors}
\usage{
calculate_hamming_distance(x, y)
}
\arguments{
\item{x}{A character vector where each element is of length 1.}

\item{y}{A character vector of the same length as x where each element is of length 1.}
}
\value{
Hamming distance as an integer.
}
\description{
Calculates Hamming distance between two vectors
}
