% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_kmer_multivalencies}
\alias{calculate_kmer_multivalencies}
\title{Calculates k-mer multivalencies}
\usage{
calculate_kmer_multivalencies(
  input_seq,
  k_len,
  window_size,
  hamming_distances,
  positional_distances
)
}
\arguments{
\item{input_seq}{sequence string}

\item{k_len}{an integer specifying the length of the k-mer}

\item{window_size}{integer specifying window_size}

\item{hamming_distances}{the Hamming distance matrix}

\item{positional_distances}{the positional distance vector}
}
\value{
a vector of k-mer multivalencies
}
\description{
Calculates k-mer multivalencies
}
