% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_kmer_pairwise_multivalencies}
\alias{calculate_kmer_pairwise_multivalencies}
\title{Calculates k-mer multivalencies with tidy output for complete 2d matrix and a positional distance matrix}
\usage{
calculate_kmer_pairwise_multivalencies(
  input_seq,
  input_seq_name,
  k_len,
  smoothing_size,
  hamming_distances,
  positional_distances
)
}
\arguments{
\item{input_seq}{sequence string}

\item{input_seq_name}{sequence string name (e.g. transcript id)}

\item{k_len}{an integer specifying the length of the k-mer}

\item{hamming_distances}{the Hamming distance matrix}

\item{positional_distances}{the positional distance matrix}

\item{smoothing}{size integer specifying smoothingwindow_size}
}
\value{
a data frame of k-mer multivalencies summed per position
}
\description{
Calculates k-mer multivalencies with tidy output for complete 2d matrix and a positional distance matrix
}
