% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_padded_sliding_mean}
\alias{calculate_padded_sliding_mean}
\title{Calculates sliding mean and pads ends with 0}
\usage{
calculate_padded_sliding_mean(iv, ws)
}
\arguments{
\item{iv}{numeric vector}

\item{ws}{sliding window size}
}
\value{
numeric vector of sliding means padded with 0
}
\description{
Calculates sliding mean and pads ends with 0
}
