% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_sliding_mean}
\alias{calculate_sliding_mean}
\title{Calculates sliding mean}
\usage{
calculate_sliding_mean(iv, ws)
}
\arguments{
\item{iv}{numeric vector}

\item{ws}{window size}
}
\value{
numeric vector of sliding means
}
\description{
Calculates sliding mean
}
