% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hamming_distance_matrix.R
\name{create_hamming_distance_matrix}
\alias{create_hamming_distance_matrix}
\title{Creates a matrix of Hamming distances between all k-mers}
\usage{
create_hamming_distance_matrix(
  k_len,
  lambda = 1,
  unweighted = FALSE,
  scale_fun = function(x) {
     1/(1 + (x^3))
 }
)
}
\arguments{
\item{k_len}{an integer specifying the length of the k-mer}

\item{lambda}{an integer specifying the lambda for the exponential decay function. Set to NULL to use custom scale function}

\item{scale_fun}{a function specifying the scaling function. Set to NULL to return the raw distances}
}
\value{
scaled Hamming distance matrix
}
\description{
Creates a matrix of Hamming distances between all k-mers
}
