% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_hamming_distance_matrix_cpp}
\alias{create_hamming_distance_matrix_cpp}
\title{Create scoring matrix based on hamming distance for kmers}
\usage{
create_hamming_distance_matrix_cpp(
  k_len,
  lambda = 1,
  unweighted = FALSE,
  scale_fun = NULL
)
}
\arguments{
\item{k_len}{an integer specifying the length of the k-mer}

\item{lambda}{an integer specifying the lambda for the exponential decay function. Set to NULL to use custom scale function}

\item{unweighted}{logical, if TRUE returns matrix of all 1s (except N vs N = 0)}

\item{scale_fun}{R function for scaling hamming distances (default: function(x) 1/(1+x^3))}
}
\value{
NumericMatrix of k-mer similarity scores where element (i,j) represents similarity between k-mer i and k-mer j
}
\description{
Create scoring matrix based on hamming distance for kmers
}
