% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_positional_distance_matrix.R
\name{create_positional_distance_matrix}
\alias{create_positional_distance_matrix}
\title{Creates positional distance matrix based on window size}
\usage{
create_positional_distance_matrix(distance_matrix, window_size, k_len)
}
\arguments{
\item{window_size}{an integer specifying the window size}

\item{k_len}{an integer specifying the length of the k-mer}
}
\value{
scaled distance integer matrix
}
\description{
Creates positional distance matrix based on window size
}
