% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_positional_distance_vector.R
\name{create_positional_distance_vector}
\alias{create_positional_distance_vector}
\title{Creates positional distance vector based on window size}
\usage{
create_positional_distance_vector(window_size, k_len)
}
\arguments{
\item{window_size}{an integer specifying the window size}

\item{k_len}{an integer specifying the length of the k-mer}
}
\value{
scaled distance integer vector
}
\description{
Creates positional distance vector based on window size
}
