% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmer_chopper}
\alias{kmer_chopper}
\title{Chops sequence into k-mers}
\usage{
kmer_chopper(input_seq, k_len)
}
\arguments{
\item{input_seq}{A character string}

\item{k_len}{an integer specifying the length of the k-mer}
}
\value{
character vector of k-mers
}
\description{
Chops sequence into k-mers
}
