% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_kmer_multivalencies}
\alias{list_kmer_multivalencies}
\title{Get list of k-mer multivalencies}
\usage{
list_kmer_multivalencies(
  ins,
  k_len,
  window_size,
  hamming_distances,
  positional_distances
)
}
\arguments{
\item{ins}{list of ??}

\item{k_len}{an integer specifying the length of the k-mer}

\item{window_size}{integer specifying window_size}

\item{hamming_distances}{the Hamming distance matrix}

\item{positional_distances}{the positional distance vector}
}
\value{
a list of k-mer multivalencies
}
\description{
Get list of k-mer multivalencies
}
