% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{list_sliding_means}
\alias{list_sliding_means}
\title{Creates a list of sliding means.}
\usage{
list_sliding_means(ins, window_size)
}
\arguments{
\item{ins}{A list of numeric vectors (intended to be from  the list_kmer_multivalencies function).}

\item{window_size}{An integer specifying the size of the window within which to calculate mean values.}
}
\value{
list of numeric vectors containing the sliding window mean values.
}
\description{
Creates a list of sliding means.
}
