% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kmer_multivalency_pdm.R
\name{plot_kmer_multivalency_pdm}
\alias{plot_kmer_multivalency_pdm}
\title{Plots smoothed k-mer multivalency for a transcript}
\usage{
plot_kmer_multivalency_pdm(
  kmer_multivalency.dt,
  k_len,
  seq,
  seq_name,
  outdir,
  interactive = FALSE,
  annotate_max = FALSE
)
}
\arguments{
\item{kmer_multivalency.dt}{k-mer multivalency dataframe/data.table}

\item{k_len}{k-mer length}

\item{seq}{Character vector of sequences from FASTA}

\item{seq_name}{Sequence name}

\item{outdir}{Output plot directory}

\item{interactive}{If TRUE, will also plot to Rstudio window}
}
\value{
PDF of smoothed k-mer multivalency plot
}
\description{
Plots smoothed k-mer multivalency for a transcript
}
