% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kmer_multivalency_pdm_heatmap.R
\name{plot_kmer_multivalency_pdm_heatmap}
\alias{plot_kmer_multivalency_pdm_heatmap}
\title{Plots k-mer multivalency matrix as heatmap with mean/max split by diagonal}
\usage{
plot_kmer_multivalency_pdm_heatmap(
  kmer_multivalency_all,
  seq_name,
  k_len,
  outdir,
  interactive = FALSE,
  max_size = 2000
)
}
\arguments{
\item{kmer_multivalency_all}{List output from calculate_kmer_pairwise_multivalencies}

\item{seq_name}{Sequence name to plot}

\item{k_len}{K-mer length}

\item{outdir}{Output plot directory}

\item{interactive}{If TRUE, will also plot to Rstudio window}

\item{max_size}{Maximum matrix size to plot before binning}
}
\value{
PNG of k-mer multivalency heatmap (if binned - mean upper, max lower)
}
\description{
Plots k-mer multivalency matrix as heatmap with mean/max split by diagonal
}
